requirements.txt
sanic==21.12.1
mysql-connector-python==8.0.28
jsonschema==4.4.0
PyJWT==2.3.0
sanic-openapi==21.12.0
sanic-cors==1.0.1
jsonschema==4.4.0
termcolor==1.1.0